Move root container to root dir
